% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_analysis.R
\name{print.cost_polypharmacy_analysis}
\alias{print.cost_polypharmacy_analysis}
\title{Print method for cost polypharmacy analysis results}
\usage{
\method{print}{cost_polypharmacy_analysis}(x, ...)
}
\arguments{
\item{x}{A cost_polypharmacy_analysis object}

\item{...}{Additional arguments (unused)}
}
\description{
Print method for cost polypharmacy analysis results
}
