% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_analysis.R
\name{summary.cost_polypharmacy_analysis}
\alias{summary.cost_polypharmacy_analysis}
\title{Summary method for cost polypharmacy analysis results}
\usage{
\method{summary}{cost_polypharmacy_analysis}(object, ...)
}
\arguments{
\item{object}{A cost_polypharmacy_analysis object}

\item{...}{Additional arguments (unused)}
}
\description{
Summary method for cost polypharmacy analysis results
}
